# Prevent hashmap growing too large. Improves performance for too many Unique Keys cases.