--- a/sdks/java/harness/src/main/java/org/apache/beam/fn/harness/PrecombineGroupingTable.java
+++ b/sdks/java/harness/src/main/java/org/apache/beam/fn/harness/PrecombineGroupingTable.java
@@ -130,6 +130,11 @@ public class PrecombineGroupingTable<K, InputT, AccumT>
   private final boolean isGloballyWindowed;
   private long lastWeightForFlush;
 
+  // Prevent hashmap growing too large. Improves performance for too many Unique Keys cases.
+  // Keep it less than (2^14)*loadFactor=(2^14)*0.75=12288
+  // Note: (2^13)*0.75=6144 looks too small to consider as limit
+  private static final int DEFAULT_MAX_GROUPING_TABLE_SIZE = 12_000;
+
   private static final class Key implements Weighted {
     private static final Key INSTANCE = new Key();
 
@@ -406,7 +411,10 @@ public class PrecombineGroupingTable<K, InputT, AccumT>
           return tableEntry;
         });
 
-    if (Caches.shouldUpdateOnSizeChange(lastWeightForFlush, weight)) {
+    if (lruMap.size() >= DEFAULT_MAX_GROUPING_TABLE_SIZE) {
+      flush(receiver);
+      lastWeightForFlush = weight;
+    } else if (Caches.shouldUpdateOnSizeChange(lastWeightForFlush, weight)) {
       flushIfNeeded(receiver);
       lastWeightForFlush = weight;
     }