@kennknowles 2023-04-19
I imagine later this value may be better dynamic, but this is a good improvement.